+++ ReleaseNotes.md
@@ -1,6 +1,6 @@
 0.2.0 - 2020-09-01
 ==================
-- Supported new command `whoami`.
+- Supported new commands `chained` and `whoami`.
 - Fixed loading previously selected role no longer exists.
 
 0.1.0 - 2019-09-30